+++ src/main/java/it/be/progettosettimana12/controller/CategoriaController.java
@@ -2,6 +2,7 @@ package it.be.progettosettimana12.controller;
 
 import java.util.List;
 import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -17,7 +18,9 @@ import org.springframework.web.bind.annotation.RestController;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import it.be.progettosettimana12.model.Categoria;
+import it.be.progettosettimana12.model.Libro;
 import it.be.progettosettimana12.service.CategoriaService;
+import it.be.progettosettimana12.service.LibroService;
 
 @SecurityRequirement(name = "bearerAuth")
 @RestController
@@ -27,6 +30,9 @@ public class CategoriaController {
 
 	@Autowired
 	private CategoriaService categoriaService;
+
+	@Autowired
+	private LibroService libroService;
 
 	@GetMapping("/findallCategorie")
 	@PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_USER')")
@@ -66,8 +72,19 @@ public class CategoriaController {
 	@PreAuthorize("hasRole('ROLE_ADMIN')")
 	@Operation(summary = "Cancella una categoria", description = "Inserire un ID di una categoria da cancellare")
 	public ResponseEntity<String> deleteCategoriaById(@PathVariable Long id) {
-		categoriaService.deleteCategoriaById(id);
-		return new ResponseEntity<>("Categoria cancellata", HttpStatus.OK);
+		Optional<Categoria> find = categoriaService.findCategoriaById(id);
+		if (!find.isEmpty()) {
+			Categoria delete = find.get();
+			List<Libro> allLibri = libroService.findAllLibri();
+			for (Libro libro : allLibri) {
+				libro.deleteAllFromSet(delete);
+			}
+			categoriaService.deleteCategoriaById(id);
+			return new ResponseEntity<>("Categoria cancellata", HttpStatus.OK);
+		} 
+		else {
+			return new ResponseEntity<>("Categoria non trovata", HttpStatus.BAD_REQUEST);
+		}
 	}
 
 	@PutMapping("/updateCategoria/{id}")
